Add crypto builder endpoint and update watchdog
